Import using submodules so can run both tests and the program from base project dir. Renamed main.py to __main__.py so it will be ran from `python -m src` Tests can be ran by using `python -m unittest` from base project directory Single tests can be ran using unittest command line interface https://docs.python.org/3/library/unittest.html#command-line-interface
